refacror: rename `BuilderArgumentInfo` to `ParseParameterInfo`
